refactor : 주소 타입 변경, 주소 삭제 예외 추가 및 리팩토링

+++ src/main/java/com/example/minzok/addresss/service/AddressServiceImpl.java
@@ -5,6 +5,7 @@ import com.example.minzok.addresss.dto.AddressResponseDto;
 import com.example.minzok.addresss.entity.Address;
 import com.example.minzok.addresss.enums.AddressType;
 import com.example.minzok.addresss.repository.AddressRepository;
+import com.example.minzok.global.error.CustomNullPointerException;
 import com.example.minzok.global.error.CustomRuntimeException;
 import com.example.minzok.global.error.ExceptionCode;
 import com.example.minzok.global.jwt.MyUserDetail;
@@ -37,7 +38,7 @@ public class AddressServiceImpl implements AddressService {
         int addressCount = addressRepository.countAddressByWithdrawnIsFalseAndMember_Email(myUserDetail.getUsername());
 
         if(addressCount == 5){
-            throw new CustomRuntimeException(ExceptionCode.ADDRESS_OVER);
+            throw new CustomRuntimeException(ExceptionCode.ADDRESS_MAX_EXCEEDED);
         }
 
         Member member = memberRepository.findMemberByEmailOrElseThrow(myUserDetail.getUsername());
@@ -50,35 +51,58 @@ public class AddressServiceImpl implements AddressService {
 
     }
 
+    /**
+     * 입력된 유저 정보로 주소 리스트 조회
+     * @param myUserDetail
+     * @return
+     */
     @Override
     public List<AddressResponseDto> findAddressByMember(MyUserDetail myUserDetail) {
         return addressRepository.findAddressByMember_Email(myUserDetail.getUsername()).stream().map(AddressResponseDto::toDto).toList();
     }
 
+    /**
+     * 주소 타입을 전부 일반 주소로 변경하고 DB에 반영
+     * 그 중 선택된 주소륻 대표 주소로 변경
+     * @param id
+     * @param myUserDetail
+     * @return
+     */
     @Transactional
     @Override
     public List<AddressResponseDto> updateAddressType(Long id, MyUserDetail myUserDetail) {
 
-
-
         List<Address> addressList = addressRepository.findAddressByMember_Email(myUserDetail.getUsername());
 
+        Address selected = addressList.stream().filter(a -> a.getId().equals(id)).findFirst().orElseThrow(
+                () -> new CustomNullPointerException(ExceptionCode.CANT_FIND_ADDRESS));
+
         for (Address address : addressList) {
-            if (address.getId().equals(id)) {
-                address.updateAddressType(AddressType.DEFAULT);
-            } else {
-                address.updateAddressType(AddressType.NORMAL);
-            }
+            address.updateAddressType(AddressType.NORMAL);
         }
+
+        addressRepository.flush();
+
+        selected.updateAddressType(AddressType.DEFAULT);
 
         return addressList.stream().map(AddressResponseDto::toDto).toList();
     }
 
     @Transactional
     @Override
-    public void deleteAddress(Long id) {
+    public void deleteAddress(Long id, MyUserDetail myUserDetail) {
+
+        int addressCount = addressRepository.countAddressByWithdrawnIsFalseAndMember_Email(myUserDetail.getUsername());
+
+        if(addressCount==1){
+            throw new CustomRuntimeException(ExceptionCode.ADDRESS_MIN_EXCEEDED);
+        }
 
         Address address = addressRepository.findAddressByIdOrElseThrow(id);
+
+        if(address.getAddressType().equals(AddressType.DEFAULT)){
+            throw new CustomRuntimeException(ExceptionCode.ADDRESS_DEFAULT_NOT_DELETED);
+        }
 
         addressRepository.delete(address);
     }
